+++ lib/presentation/pages/audio/sub/songs/widgets/song_tile_more_button_widget.dart
@@ -1,7 +1,6 @@
 import 'package:color_log/color_log.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hugeicons/hugeicons.dart';
@@ -10,7 +9,7 @@ import 'package:open_player/data/services/favorite_audio_hive_service/audio_hive
 import 'package:open_player/data/services/file_service/file_service.dart';
 import 'package:open_player/logic/audio_bloc/audios_bloc.dart';
 import 'package:open_player/presentation/common/widgets/audio_info_sheet/audio_info_sheet_widget.dart';
-import 'package:open_player/presentation/pages/audio/sub/playlists/view/audio_playlists_page.dart';
+import 'package:open_player/presentation/common/widgets/create_new_playlist_button.dart';
 import 'package:open_player/utils/app_menu.dart';
 import 'package:share_plus/share_plus.dart';
 import 'package:velocity_x/velocity_x.dart';
@@ -277,20 +276,4 @@ PopupMenuItem<dynamic> _more(BuildContext context, AudioModel audio) {
       trailing: Icon(HugeIcons.strokeRoundedInformationCircle),
     ),
   );
-}
-
-class CreateNewPlaylistButton extends HookWidget {
-  const CreateNewPlaylistButton({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final ValueNotifier<String> playlistName = useState("");
-    return ElevatedButton.icon(
-      onPressed: () {
-        showCreateAudioPlaylistDialog(context, playlistName);
-      },
-      icon: Icon(HugeIcons.strokeRoundedPlaylist01),
-      label: Text("Create New Playlist"),
-    );
-  }
 }
